Create boardSolved.py
Simple python code to check if a Tic-Tac-Toe game has been solved by either X or O. Can implement in the generation of all possible boards to see if a board is completed, and if it has been completed, we want to exclude this from our list. This list will be used in training our AI model.